refactor: Consolidate duplicated board loading logic into Core class
- Created shared loadBoardTasks() method in Core class
- Removed duplicate task loading logic from cli.ts and view-switcher.ts
- Both CLI and UI now use the same core logic for loading board data
- Maintains all previous functionality including checkActiveBranches config support
- Reduces code duplication by ~80 lines

This refactoring ensures consistency between CLI and UI board views and makes
future maintenance easier by having a single source of truth for board loading logic.
