+++ src/ui/view-switcher.ts
@@ -118,86 +118,8 @@ class BackgroundLoader {
 				throw new Error("Loading cancelled");
 			}
 
-			// Import these dynamically to avoid circular deps
-			const { loadRemoteTasks, resolveTaskConflict, getTaskLoadingMessage } = await import("../core/remote-tasks.ts");
-			const { filterTasksByLatestState, getLatestTaskStatesForIds } = await import("../core/cross-branch-tasks.ts");
-
-			const config = await this.core.filesystem.loadConfig();
-			const statuses = config?.statuses || [];
-			const resolutionStrategy = config?.taskResolutionStrategy || "most_progressed";
-
-			// Check for cancellation before loading
-			if (this.abortController?.signal.aborted) {
-				throw new Error("Loading cancelled");
-			}
-
-			// Load local and remote tasks in parallel
-			this.onProgress?.(getTaskLoadingMessage(config));
-			const [localTasks, remoteTasks] = await Promise.all([
-				this.core.listTasksWithMetadata(),
-				loadRemoteTasks(this.core.gitOps, config, this.onProgress),
-			]);
-
-			// Check for cancellation after loading basic tasks
-			if (this.abortController?.signal.aborted) {
-				throw new Error("Loading cancelled");
-			}
-
-			// Create map with local tasks
-			const tasksById = new Map<string, Task>(localTasks.map((t) => [t.id, { ...t, source: "local" }]));
-
-			// Merge remote tasks with local tasks
-			this.onProgress?.("Resolving task states across branches...");
-			for (const remoteTask of remoteTasks) {
-				// Check for cancellation during merge
-				if (this.abortController?.signal.aborted) {
-					throw new Error("Loading cancelled");
-				}
-
-				const existing = tasksById.get(remoteTask.id);
-				if (!existing) {
-					tasksById.set(remoteTask.id, remoteTask);
-				} else {
-					const resolved = resolveTaskConflict(existing, remoteTask, statuses, resolutionStrategy);
-					tasksById.set(remoteTask.id, resolved);
-				}
-			}
-
-			// Check for cancellation before final steps
-			if (this.abortController?.signal.aborted) {
-				throw new Error("Loading cancelled");
-			}
-
-			// Get the latest directory location of each task across all branches
-			const tasks = Array.from(tasksById.values());
-			let filteredTasks: Task[];
-
-			if (config?.checkActiveBranches === false) {
-				// Skip cross-branch checking for maximum performance
-				this.onProgress?.("Skipping cross-branch check (disabled in config)...");
-				filteredTasks = tasks;
-			} else {
-				const taskIds = tasks.map((t) => t.id);
-				const latestTaskDirectories = await getLatestTaskStatesForIds(
-					this.core.gitOps,
-					this.core.filesystem,
-					taskIds,
-					this.onProgress,
-					{
-						recentBranchesOnly: true,
-						daysAgo: config?.activeBranchDays ?? 30,
-					},
-				);
-
-				// Check for cancellation before filtering
-				if (this.abortController?.signal.aborted) {
-					throw new Error("Loading cancelled");
-				}
-
-				// Filter tasks based on their latest directory location
-				this.onProgress?.("Filtering active tasks...");
-				filteredTasks = filterTasksByLatestState(tasks, latestTaskDirectories);
-			}
+			// Use the shared Core method for loading board tasks
+			const filteredTasks = await this.core.loadBoardTasks(this.onProgress, this.abortController?.signal);
 
 			// Cache the results
 			this.cachedTasks = filteredTasks;
